refactored admin test folder to be easier to navigate tests for different functions
