+++ frontend/src/__tests__/admin/edit/handleSaveDiscipline.test.js
@@ -1,4 +1,4 @@
-import { handleSaveDiscipline } from '../../utils/adminFetching'
+import { handleSaveDiscipline } from '../../../utils/adminFetching'
 
 global.fetch = jest.fn()
