+++ frontend/src/__tests__/admin/edit/handleSaveMajor.test.js
@@ -1,4 +1,4 @@
-import { handleSaveMajor } from '../../utils/adminFetching'
+import { handleSaveMajor } from '../../../utils/adminFetching'
 
 global.fetch = jest.fn()
