+++ frontend/src/__tests__/admin/edit/handleSaveUmbrella.test.js
@@ -1,4 +1,4 @@
-import { handleSaveUmbrella } from '../../utils/adminFetching'
+import { handleSaveUmbrella } from '../../../utils/adminFetching'
 
 global.fetch = jest.fn()
