+++ frontend/src/__tests__/admin/view/AdminFacultyEdit.test.js
@@ -3,8 +3,8 @@ import { render, screen, waitFor, fireEvent } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { MemoryRouter, useNavigate, useParams } from 'react-router-dom'
 import { ThemeProvider, createTheme } from '@mui/material/styles'
-import AdminFacultyEdit from '../../components/admin/AdminFacultyEdit'
-import { getDepartmentsExpectedResponse, getFacultyExpectedResponse, putFacultyExpectedRequest } from '../../resources/mockData'
+import AdminFacultyEdit from '../../../components/admin/AdminFacultyEdit'
+import { getDepartmentsExpectedResponse, getFacultyExpectedResponse, putFacultyExpectedRequest } from '../../../resources/mockData'
 
 // Wrap component with ThemeProvider and MemoryRouter
 const renderWithTheme = (ui) => {
