+++ frontend/src/__tests__/admin/view/renderResearchPeriods.test.js
@@ -1,6 +1,6 @@
 import { render, screen, fireEvent } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
-import { renderResearchPeriods } from '../../components/admin/RenderAdminVariables'
+import { renderResearchPeriods } from '../../../components/admin/RenderAdminVariables'
 
 describe('renderResearchPeriods Component', () => {
   const defaultProps = {
